💂🤘 Checkpoint ./cmd/addnabor.js:50975678/3694 ./cmd/aunwarn.js:50975678/5431 ./cmd/awarn.js:50975678/8625 ./cmd/checkbd.js:50975678/1985 ./cmd/dadd.js:50975678/3725 ./cmd/ddown.js:50975678/4981 ./cmd/disable.js:50975678/1168 ./cmd/down.js:50975678/8721 ./cmd/dremove.js:50975678/4313 ./cmd/dup.js:50975678/4957 ./cmd/enable.js:50975678/1082 ./cmd/madd.js:50975678/3989 ./cmd/mdown.js:50975678/12937 ./cmd/mremove.js:50975678/5519 ./cmd/reboot.js:50975678/683 ./cmd/mup.js:50975678/12939 ./cmd/nstart.js:50975678/2539 ./cmd/nstop.js:50975678/2269 ./cmd/prinyat.js:50975678/3831 ./cmd/removenabor.js:50975678/3593 ./cmd/rpunwarn.js:50975678/6309 ./cmd/rpwarn.js:50975678/8289 ./cmd/say.js:50975678/2259 ./cmd/setbd.js:50975678/2285 ./cmd/up.js:50975678/8713 ./cmd/uval.js:50975678/5833 ./cmd/vadd.js:50975678/4003 ./cmd/vdown.js:50975678/12765 ./cmd/vremove.js:50975678/5367 ./cmd/vup.js:50975678/12767 ./index.js:50975678/2785

+++ cmd/addnabor.js
@@ -5,9 +5,11 @@ module.exports = {
     name: 'набор-выдать',
     description: '',
     run: async(client, message, args) => {
+        let disable = await db.get(`disable`);
+        if(disable === 1) return
+        client.channels.cache.get('1043770641867882546').send(`${message.author} ввёл команду: ${message}`);
         message.channel.bulkDelete(1);
         await message.guild.roles.fetch()
-        client.channels.cache.get('1043770641867882546').send(`${message.author} ввёл команду: ${message}`);
         if(message.member.permissions.has("ADMINISTRATOR"));
         else if(!message.member.roles.cache.has("1040570760982441984"));
         else if(!message.member.roles.cache.has("1040570130314309656")) return;
