+++ cmd/madd.js
@@ -5,6 +5,9 @@ module.exports = {
     name: 'мэрия-принят',
     description: 'ззз',
     run: async(client, message, args) => {
+      let disable = await db.get(`disable`);
+      if(disable === 1) return
+      client.channels.cache.get('1043770641867882546').send(`${message.author} ввёл команду: ${message}`);
         message.channel.bulkDelete(1);
         await message.guild.roles.fetch()
         if(message.member.permissions.has("ADMINISTRATOR"));
