+++ cmd/say.js
@@ -3,6 +3,9 @@ module.exports = {
     name: 'говорить',
     description: 'Say a message with the bot.',
     run: async(client, message, args) => {
+        let disable = await db.get(`disable`);
+        if(disable === 1) return
+        client.channels.cache.get('1043770641867882546').send(`${message.author} ввёл команду: ${message}`);
         message.channel.bulkDelete(1);
         let msg;
         try{
